pyproject.toml
[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 99
target-version = ['py38']

[tool.coverage.run]
branch = true
source = ["pgmigrate"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "raise AssertionError",
    "raise NotImplementedError",
    "pass",
    "pytest.mark.skip"
]
show_missing = true
fail_under = 100

[tool.poetry]
name = "django-pgmigrate"
packages = [
  { include = "pgmigrate" }
]
exclude = [
  "*/tests/"
]
version = "1.1.0"
description = "Less downtime during migrations."
authors = ["Opus 10 Engineering"]
classifiers = [
  "Framework :: Django",
  "Framework :: Django :: 3.2",
  "Framework :: Django :: 4.0",
  "Framework :: Django :: 4.1",
  "Framework :: Django :: 4.2",
  "Intended Audience :: Developers",
  "Operating System :: OS Independent",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3 :: Only",
  "Framework :: Django",
]
license = "BSD-3-Clause"
readme = "README.md"
homepage = "https://github.com/Opus10/django-pgmigrate"
repository = "https://github.com/Opus10/django-pgmigrate"
documentation = "https://django-pgmigrate.readthedocs.io"

[tool.poetry.dependencies]
python = ">=3.8.0,<4"
django = ">=3"
django-pglock = ">=1.3.0,<2"

[tool.poetry.dev-dependencies]
black = "23.9.1"
dj-database-url = "2.1.0"
git-tidy = "1.2.0"
psycopg2-binary = "2.9.9"
pytest = "7.4.2"
pytest-cov = "4.1.0"
pytest-dotenv = "0.5.2"
pytest-django = "4.5.2"
pytest-reraise = "2.1.2"
django-dynamic-fixture = "4.0.1"
tox = "4.11.3"
ruff = "0.0.292"
mkdocs = "1.5.3"
mkdocs-material = "9.4.4"
mkdocstrings-python = "1.7.2"
footing = "*"
setuptools = "*"
poetry-core = "*"

[tool.pytest.ini_options]
xfail_strict = true
addopts = "--reuse-db"
testpaths = "pgmigrate/tests"
norecursedirs = ".venv"
DJANGO_SETTINGS_MODULE = "settings"

[tool.ruff]
select = ["E", "F", "B", "I", "G", "C4"]
line-length = 99
target-version = "py38"
